feat: 采集详情 (#788)

+++ bkmonitor/webpack/src/monitor-pc/pages/alarm-dispatch/alarm-dispatch.tsx
@@ -61,6 +61,8 @@ export default class AlarmDispatch extends tsc<{}> {
   cacheRuleGroups = [];
   /* 搜索 */
   search = '';
+  /** 是否能查询，需要依赖 getAlarmGroupList 和 getAlarmDispatchGroupData两个函数都请求完成 */
+  isSearch = false;
   /** 新建规则组弹窗 */
   visible = false;
 
@@ -176,18 +178,21 @@ export default class AlarmDispatch extends tsc<{}> {
   }
 
   created() {
+    this.getRouteParams();
     this.getAlarmDispatchGroupData();
     this.getAlarmGroupList();
     this.getKVOptionsData();
     this.getProcessPackage();
     this.$store.commit('app/SET_NAV_ROUTE_LIST', [{ name: this.$t('route-告警分派'), id: '' }]);
-    this.getRouteParams();
   }
 
   getRouteParams() {
-    const { groupName } = this.$route.params;
+    const { groupName, group_id: groupId } = this.$route.query;
     if (groupName) {
-      this.search = groupName;
+      this.search = groupName as string;
+    }
+    if (groupId) {
+      this.search = groupId as string;
     }
   }
 
@@ -199,6 +204,12 @@ export default class AlarmDispatch extends tsc<{}> {
       name: item.name,
       receiver: item.users?.map(rec => rec.display_name) || []
     }));
+
+    /** 能否查询需要依赖 getAlarmGroupList 和 getAlarmDispatchGroupData两个函数都请求完成 */
+    if (this.isSearch && this.search) {
+      this.handleSearch();
+    }
+    this.isSearch = true;
   }
 
   /**
@@ -217,7 +228,8 @@ export default class AlarmDispatch extends tsc<{}> {
             name: item.name,
             priority: item.priority,
             isExpan: true,
-            ruleData: []
+            ruleData: [],
+            editAllowed: !!item?.edit_allowed
           })
       ) || [];
     this.loading = false;
@@ -242,7 +254,10 @@ export default class AlarmDispatch extends tsc<{}> {
     });
     this.cacheRuleGroups = this.ruleGroups;
     this.loading = false;
-    this.search && this.handleSearch();
+    if (this.isSearch && this.search) {
+      this.handleSearch();
+    }
+    this.isSearch = true;
   }
 
   // 获取流程套餐
@@ -486,6 +501,12 @@ export default class AlarmDispatch extends tsc<{}> {
     } else {
       this.emptyText = window.i18n.tc('查无数据');
       this.ruleGroups = this.cacheRuleGroups;
+      const { groupName, groupId, ...arg } = this.$route.query;
+      this.$router.replace({
+        query: {
+          ...arg
+        }
+      });
     }
   }
 
@@ -591,16 +612,26 @@ export default class AlarmDispatch extends tsc<{}> {
                     {this.renderEditAttribute(item.name, item.ruleData.length, 'name', item.id)}
                     {this.renderEditAttribute(`${this.$t('优先级')}:`, item.priority, 'priority', item.id)}
                     <div
-                      class='edit-btn-wrap'
-                      onClick={() => this.handleToConfig(item.id)}
+                      class={['edit-btn-wrap', { 'edit-btn-disabled': !item.editAllowed }]}
+                      v-bk-tooltips={{
+                        placements: ['top'],
+                        content: this.$t('内置的分派规则组不允许修改'),
+                        disabled: item.editAllowed
+                      }}
+                      onClick={() => item.editAllowed && this.handleToConfig(item.id)}
                     >
                       <span class='icon-monitor icon-bianji'></span>
                       <span>{this.$t('配置规则')}</span>
                     </div>
                     <div
-                      class='del-btn-wrap'
+                      class={['del-btn-wrap', { 'del-btn-disabled': !item.editAllowed }]}
+                      v-bk-tooltips={{
+                        placements: ['top'],
+                        content: this.$t('内置的分派规则组不允许修改'),
+                        disabled: item.editAllowed
+                      }}
                       onClick={e => {
-                        this.handleDeleteGroup(e, item);
+                        item.editAllowed && this.handleDeleteGroup(e, item);
                       }}
                     >
                       <span class='icon-monitor icon-mc-delete-line'></span>
